Remove need for visualise specific config sources
