--- generator/src/main/java/com/scottlogic/deg/generator/ConfigSource.java
@@ -1,18 +1,0 @@
-package com.scottlogic.deg.generator;
-
-import java.io.File;
-import java.nio.file.Path;
-
-/**
- * Parent for command line configuration options.
- */
-public interface ConfigSource {
-
-    boolean dontOptimise();
-    Path getOutputPath();
-    File getProfileFile();
-    boolean isSchemaValidationEnabled();
-    boolean getValidateProfile();
-    boolean overwriteOutputFiles();
-
-}
