+++ generator/src/main/java/com/scottlogic/deg/generator/commandline/VisualiseCommandLine.java
@@ -1,11 +1,11 @@
 package com.scottlogic.deg.generator.commandline;
 
+import com.google.inject.Guice;
+import com.google.inject.Injector;
 import com.scottlogic.deg.generator.VisualiseExecute;
-import com.scottlogic.deg.generator.visualisation.VisualisationConfigSource;
+import com.scottlogic.deg.generator.guice.BaseModule;
 import picocli.CommandLine;
 
-import java.io.File;
-import java.nio.file.Path;
 
 /**
  * This class holds the visualisation specific command line options.
@@ -20,67 +20,15 @@ import java.nio.file.Path;
     parameterListHeading = "%nParameters:%n",
     optionListHeading = "%nOptions:%n",
     abbreviateSynopsis = true)
-public class VisualiseCommandLine extends CommandLineBase implements VisualisationConfigSource {
-
-    @CommandLine.Parameters(index = "1", description = "The path of the output visualise file.")
-    private Path outputPath;
-
-    @CommandLine.Option(
-        names = {"-t", "--title"},
-        description = "The title to place at the top of the file")
-    private String titleOverride;
-
-    @CommandLine.Option(
-        names = {"--no-title"},
-        description = "Hides the title from the output")
-    private boolean shouldHideTitle;
-
-    @CommandLine.Option(
-        names = {"--no-simplify"},
-        description = "Prevents tree simplification",
-        hidden = true)
-    private boolean dontSimplify;
+public class VisualiseCommandLine extends GenerateCommandLine  {
 
     @Override
-    public Path getOutputPath() {
-        return outputPath;
-    }
+    public void run() {
+        BaseModule container = new BaseModule(this);
+        Injector injector = Guice.createInjector(container);
 
-    @Override
-    public File getProfileFile() {
-        return profileFile;
-    }
+        Runnable task = injector.getInstance(VisualiseExecute.class);
 
-    @Override
-    public boolean isSchemaValidationEnabled() {
-        return this.enableSchemaValidation;
-    }
-
-    public String getTitleOverride() {
-        return titleOverride;
-    }
-
-    public boolean shouldHideTitle() {
-        return shouldHideTitle;
-    }
-
-    @Override
-    public boolean dontOptimise() {
-        return dontOptimise;
-    }
-
-    @Override
-    public boolean overwriteOutputFiles() {
-        return overwriteOutputFiles;
-    }
-
-    @Override
-    public boolean getValidateProfile() {
-        return false;
-    }
-
-    @Override
-    protected Class<? extends Runnable> getExecutorType() {
-        return VisualiseExecute.class;
+        task.run();
     }
 }
