+++ generator/src/main/java/com/scottlogic/deg/generator/generation/GenerationConfigSource.java
@@ -1,38 +1,33 @@
 package com.scottlogic.deg.generator.generation;
 
-import com.scottlogic.deg.generator.ConfigSource;
 import com.scottlogic.deg.generator.config.detail.*;
 import com.scottlogic.deg.profile.v0_1.AtomicConstraintType;
 
+import java.io.File;
+import java.nio.file.Path;
 import java.util.List;
 
-public interface GenerationConfigSource extends ConfigSource {
-    /**
-     * Gets a value signifying the data generation type, i.e. one of interesting, full sequential or random.
-     * @return Enum value of the current generation types.
-     */
-    DataGenerationType getGenerationType();
+public interface GenerationConfigSource  {
+    File getProfileFile();
+    boolean isSchemaValidationEnabled();
 
-    /**
-     * Gets a value signifying the combination strategy, i.e. one of exhaustive, pinning or minimal
-     * @return Enum value of the current combination strategy.
-     */
-    CombinationStrategyType getCombinationStrategyType();
-
-    /**
-     * Gets a value signifying the current tree walker type, i.e. the reductive walker.
-     * @return Enum value of the current decision tree walker.
-     */
-    TreeWalkerType getWalkerType();
-    MonitorType getMonitorType();
-    List<AtomicConstraintType> getConstraintsToNotViolate();
-    long getMaxRows();
-    boolean shouldDoPartitioning();
-    boolean isEnableTracing();
-    boolean overwriteOutputFiles();
-    boolean visualiseReductions();
     boolean shouldViolate();
-    boolean requireFieldTyping();
+    List<AtomicConstraintType> getConstraintsToNotViolate();
+
+    DataGenerationType getGenerationType();
+    CombinationStrategyType getCombinationStrategyType();
+    TreeWalkerType getWalkerType();
+    long getMaxRows();
 
     OutputFormat getOutputFormat();
+    Path getOutputPath();
+    boolean overwriteOutputFiles();
+
+
+    MonitorType getMonitorType();
+    boolean shouldDoPartitioning();
+    boolean isEnableTracing();
+    boolean visualiseReductions();
+    boolean requireFieldTyping();
+    boolean dontOptimise();
 }
