+++ generator/src/main/java/com/scottlogic/deg/generator/guice/ProfileSchemaValidatorProvider.java
@@ -2,18 +2,18 @@ package com.scottlogic.deg.generator.guice;
 
 import com.google.inject.Inject;
 import com.google.inject.Provider;
-import com.scottlogic.deg.generator.ConfigSource;
+import com.scottlogic.deg.generator.generation.GenerationConfigSource;
 import com.scottlogic.deg.profile.v0_1.NoopProfileSchemaValidator;
 import com.scottlogic.deg.profile.v0_1.ProfileSchemaValidator;
 import com.scottlogic.deg.profile.v0_1.ProfileSchemaValidatorLeadPony;
 
 public class ProfileSchemaValidatorProvider implements Provider<ProfileSchemaValidator> {
 
-    private final ConfigSource configSource;
+    private final GenerationConfigSource configSource;
     private final ProfileSchemaValidatorLeadPony leadPonyValidator;
 
     @Inject
-    public ProfileSchemaValidatorProvider(ConfigSource configSource, ProfileSchemaValidatorLeadPony leadPonyValidator) {
+    public ProfileSchemaValidatorProvider(GenerationConfigSource configSource, ProfileSchemaValidatorLeadPony leadPonyValidator) {
         this.configSource = configSource;
         this.leadPonyValidator = leadPonyValidator;
     }
