+++ generator/src/test/java/com/scottlogic/deg/generator/visualisation/TestVisualisationConfigSource.java
@@ -1,12 +1,11 @@
 package com.scottlogic.deg.generator.visualisation;
 
+import com.scottlogic.deg.generator.commandline.VisualiseCommandLine;
+
 import java.io.File;
 import java.nio.file.Path;
 
-public class TestVisualisationConfigSource implements VisualisationConfigSource {
-
-    public TestVisualisationConfigSource() {
-    }
+public class TestVisualisationConfigSource extends VisualiseCommandLine {
 
     @Override
     public Path getOutputPath() {
@@ -24,27 +23,12 @@ public class TestVisualisationConfigSource implements VisualisationConfigSource 
     }
 
     @Override
-    public String getTitleOverride() {
-        return null;
-    }
-
-    @Override
-    public boolean shouldHideTitle() {
-        return false;
-    }
-
-    @Override
     public boolean dontOptimise() {
         return false;
     }
 
     @Override
     public boolean overwriteOutputFiles() {
-        return false;
-    }
-
-    @Override
-    public boolean getValidateProfile() {
         return false;
     }
 }
